fix delete story if no flights left

+++ app/models/flight.rb
@@ -16,11 +16,10 @@ class Flight < ApplicationRecord
   around_destroy :destroy_orphaned_story
 
   def destroy_orphaned_story
-    if story = self.story
-      yield # executes a DELETE database statement
-      if story.flights.length == 0
-        story.destroy
-      end
+    story = self.story
+    yield # executes a DELETE database statement
+    if story && story.flights.length == 0
+      story.destroy
     end
   end
